requirements.txt
fastapi
uvicorn
python-multipart
rembg[cpu]
pillow
numpy
opencv-python-headless
requests
streamlit